+++ pkg/pfconfig/basicstationlns/basicstationlns.go
@@ -116,10 +116,13 @@ func GetRouterConfig(bandID string, fps map[string]*frequencyplans.FrequencyPlan
 		// These fields are not defined in the v1.5 ref design https://doc.sm.tc/station/gw_v1.5.html#rfconf-object and would cause a parsing error.
 		sx1301Conf.Radios[0].TxFreqMin = 0
 		sx1301Conf.Radios[0].TxFreqMax = 0
-		// Remove hardware specific values that are not necessary.
+		// Remove hardware specific values that are handled by the gateway itself.
+		sx1301Conf.ClockSource = 0
 		sx1301Conf.TxLUTConfigs = nil
 		for i := range sx1301Conf.Radios {
 			sx1301Conf.Radios[i].Type = ""
+			sx1301Conf.Radios[i].TxEnable = false
+			sx1301Conf.Radios[i].RSSIOffset = 0
 		}
 		if err != nil {
 			return RouterConfig{}, err