+++ CHANGELOG.md
@@ -26,6 +26,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Removed
 
+- Hardware specific fields from Basic Station router config.
+
 ### Fixed
 
 - Telemetry and events for gateway statuses.